git-svn-id: https://svn.r-project.org/R-packages/trunk/Mac-GUI@5713 694ef91d-65df-0310-b7bb-92e67a308ead

+++ PackageManager.m
@@ -64,12 +64,12 @@ static PackageManager *sharedController = nil;
 #pragma mark TableView delegates
 
 /* These two routines are needed to update the History TableView */
-- (int)numberOfRowsInTableView: (NSTableView *)tableView
+- (NSInteger)numberOfRowsInTableView: (NSTableView *)tableView
 {
 	return packages;
 }
 
-- (id)tableView: (NSTableView *)tableView objectValueForTableColumn: (NSTableColumn *)tableColumn row: (int)row
+- (id)tableView: (NSTableView *)tableView objectValueForTableColumn: (NSTableColumn *)tableColumn row: (NSInteger)row
 {
 	if (row<packages) {
 		if([[tableColumn identifier] isEqualToString:@"status"])
@@ -85,7 +85,7 @@ static PackageManager *sharedController = nil;
 - (void)tableView:(NSTableView *)tableView
 	setObjectValue:(id)object
 	forTableColumn:(NSTableColumn *)tableColumn
-	row:(int)row
+	row:(NSInteger)row
 {
 	if (row>=packages) return;
 	if([[tableColumn identifier] isEqualToString:@"status"]){
